+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2022-07-29
+
+### Added
+
+-   getCSRF documentation in the README
+-   getCSRF function
+-   speedtest function
+-   deleteAccount function
+-   disableAccount function
+-   restoreAccount function
+
+### Changed
+
+-   mac appid example
+
 ## [2.1.0] - 2022-06-30
 
 ### Added